Advanced Software Dev (CS 3240)/Course project/requirements-travis.txt
Django
gunicorn
django-heroku
psycopg2-binary
dj-database-url
social-auth-app-django
django-crispy-forms